requirements.txt

folium==0.18.0
inflection==0.5.1
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
streamlit==1.40.1
streamlit_folium==0.23.2